fix watching for imports of discovered MML documents
This fixes an issue where imports of discovered documents were not being
watched by esbuild and therefore would not trigger a rebuild when
changed. This moves the rebuild logic up a a level so we are always
watching the process with all discovered entrypoints, rather than just
the _first_ process as it was previously.

The best fix for this is to have esbuild support dynamically adding
entry points or to make the onDispose hook synchronous so we don't have
to but the `waitForDispose` hack in the tests.